legal-agent/requirements.txt
groq>=0.9.0
sentence-transformers>=3.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
python-dotenv>=1.0.0
streamlit>=1.35.0
huggingface-hub>=0.23.0
colorama>=0.4.6
